refactor: remove unused Stripe API services and related types
