+++ src/lib/types/todos.ts
@@ -1,7 +1,14 @@
-export type TodoT = {
-  id?: string;
-  title: string;
-  description: string;
-  author?: string;
-  user_id?: string;
-};
+export interface Task {
+  id: string;
+  goalId: string;
+  text: string;
+  isCompleted: boolean;
+  lastCompleted?: string | null;
+}
+
+export interface Goal {
+  id: string;
+  name: string;
+  tasks: Task[];
+  created_at: string;
+}
